requirements.txt
Flask==3.1.0
pymongo==4.10.1
paho-mqtt==1.5.1
flask-cors==5.0.0
Werkzeug==3.1.3
python-dotenv==0.19.2
gunicorn==23.0.0
dnspython==2.7.0